+++ src/settings/MappingProfiles/detailsSections/edit/InvoiceDetailSection/tests/InvoiceLineAdjustments.test.js
@@ -1,0 +1,70 @@
+import React from 'react';
+import { fireEvent } from '@testing-library/react';
+import {
+  noop,
+  get,
+} from 'lodash';
+
+import { renderWithIntl } from '@folio/stripes-data-transfer-components/test/jest/helpers';
+
+import '../../../../../../../test/jest/__mock__';
+import {
+  renderWithReduxForm,
+  translationsProperties,
+} from '../../../../../../../test/jest/helpers';
+
+import { InvoiceLineAdjustments } from '../InvoiceLineAdjustments';
+import INVOICE from '../../../../initialDetails/INVOICE';
+
+const mockLineAdjustmentsFields = get(INVOICE.mappingFields, 'invoiceLines.[0].fields[15].subfields[0]');
+
+jest.mock('../../../hooks', () => ({
+  useFieldMappingFieldValue: () => ['USD'],
+  useFieldMappingRefValues: () => [[mockLineAdjustmentsFields]],
+}));
+
+const renderInvoiceLineAdjustments = () => {
+  const component = () => (
+    <InvoiceLineAdjustments
+      invoiceLinesFieldIndex={26}
+      setReferenceTables={noop}
+      mappingFields={INVOICE.mappingFields}
+      initialFields={{}}
+      okapi={{}}
+    />
+  );
+
+  return renderWithIntl(renderWithReduxForm(component), translationsProperties);
+};
+
+describe('InvoiceLineAdjustments', () => {
+  it('should render correct fields', () => {
+    const { getByRole } = renderInvoiceLineAdjustments();
+
+    expect(getByRole('button', { name: 'Add adjustment' })).toBeInTheDocument();
+  });
+
+  describe('when clicking "Add adjustment" button', () => {
+    it('new Adjustment card should be rendered', () => {
+      const {
+        getByRole,
+        getByText,
+      } = renderInvoiceLineAdjustments();
+
+      const addButton = getByRole('button', { name: 'Add adjustment' });
+      fireEvent.click(addButton);
+
+      expect(getByText('Adjustment 1')).toBeInTheDocument();
+    });
+
+    it('with correct fields', () => {
+      const { getByText } = renderInvoiceLineAdjustments();
+
+      expect(getByText('Description')).toBeInTheDocument();
+      expect(getByText('Amount')).toBeInTheDocument();
+      expect(getByText('Type')).toBeInTheDocument();
+      expect(getByText('Relation to total')).toBeInTheDocument();
+      expect(getByText('Export to accounting')).toBeInTheDocument();
+    });
+  });
+});